+++ spec/integration/resource/api300/synergy/sas_logical_interconnect/delete_spec.rb
@@ -1,0 +1,7 @@
+require 'spec_helper'
+
+klass = OneviewSDK::API300::Synergy::SASLogicalInterconnect
+RSpec.describe klass, integration: true, type: DELETE, sequence: rseq(klass) do
+  let(:current_client) { $client_300_synergy }
+  include_examples 'SASLogicalInterconnectDeleteExample', 'integration api300 context'
+end